--- lib/css/styles.css
@@ -1,8 +1,0 @@
-body {
-    height: 100%;
-}
-
-.container {
-    display: flex;
-    flex-direction: column;
-}
