--- logo.html
@@ -1,28 +1,0 @@
-
-    <!DOCTYPE html>
-    <html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <link rel="stylesheet" href="./css/styles.css">
-
-        <title>SVG Logo</title>
-    </head>
-    <body>
-        <header class='container'>
-        <h1>Here is your logo</h1>
-        </header>
-
-
-        <div clas='container'>
-        <svg width="100" height="100">
-            <rect width="400" height="100" 
-            style="fill:black;stroke-width:10;stroke:rgb(105, 75, 88)" />
-            <text x="50" y="50" fill="green">FER</text>
-        </svg>
-        </div>
-    </body>
-    </html>
-    
-
